Module uniffi_bindgen::backend

Structs

TemplateExpression

Config value for template expressions

Traits

CodeDeclaration

A trait that is able to render a declaration about a particular member declared in the ComponentInterface. Like CodeType, it can render declaration code and imports. It also is able to render code at start-up of the FFI. All methods are optional, and there is no requirement that the trait be used for a particular interface:: member. Thus, it can also be useful for conditionally rendering code.

CodeOracle

An object to look up a foreign language code specific renderer for a given type used. Every Type referred to in the ComponentInterface should map to a corresponding CodeType.

CodeType

A Trait to emit foreign language code to handle referenced types. A type which is specified in the UDL (i.e. a member of the component interface) will have a CodeDeclaration as well, but for types used e.g. primitive types, Strings, etc only a CodeType is needed.

Type Definitions

Literal

TypeIdentifier