Enum uniffi_bindgen::backend::Type

pub enum Type {

    UInt8,
    Int8,
    UInt16,
    Int16,
    UInt32,
    Int32,
    UInt64,
    Int64,
    Float32,
    Float64,
    Boolean,
    String,
    Bytes,
    Timestamp,
    Duration,
    Object {
        module_path: String,
        name: String,
        imp: ObjectImpl,
    },
    Record {
        module_path: String,
        name: String,
    },
    Enum {
        module_path: String,
        name: String,
    },
    CallbackInterface {
        module_path: String,
        name: String,
    },
    Optional {
        inner_type: Box<Type, Global>,
    },
    Sequence {
        inner_type: Box<Type, Global>,
    },
    Map {
        key_type: Box<Type, Global>,
        value_type: Box<Type, Global>,
    },
    External {
        module_path: String,
        name: String,
        namespace: String,
        kind: ExternalKind,
        tagged: bool,
    },
    Custom {
        module_path: String,
        name: String,
        builtin: Box<Type, Global>,
    },

}

Represents all the different high-level types that can be used in a component interface. At this level we identify user-defined types by name, without knowing any details of their internal structure apart from what type of thing they are (record, enum, etc).

Variants

UInt8

Int8

UInt16

Int16

UInt32

Int32

UInt64

Int64

Float32

Float64

Boolean

String

Bytes

Timestamp

Duration

Object

Fields

module_path: String

name: String

imp: ObjectImpl

Record

Fields

module_path: String

name: String

Enum

Fields

module_path: String

name: String

CallbackInterface

Fields

module_path: String

name: String

Optional

Fields

inner_type: Box<Type, Global>

Sequence

Fields

inner_type: Box<Type, Global>

Map

Fields

key_type: Box<Type, Global>

value_type: Box<Type, Global>

External

Fields

module_path: String

name: String

namespace: String

kind: ExternalKind

tagged: bool

Custom

Fields

module_path: String

name: String

builtin: Box<Type, Global>

Implementations

impl Type

pub fn iter_types(&self) -> Box<dyn Iterator<Item = &Type> + '_, Global>