Trait uniffi_bindgen::backend::CodeDeclaration

pub trait CodeDeclaration {
    fn imports(&self, _oracle: &dyn CodeOracle) -> Option<Vec<String>> { ... }

    fn initialization_code(&self, _oracle: &dyn CodeOracle) -> Option<String> { ... }

    fn definition_code(&self, _oracle: &dyn CodeOracle) -> Option<String> { ... }
}

A trait that is able to render a declaration about a particular member declared in the ComponentInterface. Like CodeType, it can render declaration code and imports. It also is able to render code at start-up of the FFI. All methods are optional, and there is no requirement that the trait be used for a particular interface:: member. Thus, it can also be useful for conditionally rendering code.

Provided methods

fn imports(&self, _oracle: &dyn CodeOracle) -> Option<Vec<String>>

A list of imports that are needed if this type is in use. Classes are imported exactly once.

fn initialization_code(&self, _oracle: &dyn CodeOracle) -> Option<String>

Code (one or more statements) that is run on start-up of the library, but before the client code has access to it.

fn definition_code(&self, _oracle: &dyn CodeOracle) -> Option<String>

Code which represents this member. e.g. the foreign language class definition for a given Object type.