Module uniffi_bindgen::bindings

Generate foreign language bindings for a uniffi component.

This module contains all the code for generating foreign language bindings, along with some helpers for executing foreign language scripts or tests.

Modules

kotlin 🔒

python 🔒

ruby 🔒

swift 🔒

Swift bindings backend for UniFFI

Structs

KotlinBindingGenerator

PythonBindingGenerator

RubyBindingGenerator

RunScriptOptions

Mode for the run_script function defined for each language

SwiftBindingGenerator

Enums

TargetLanguage

Enumeration of all foreign language targets currently supported by this crate.

Functions

kotlin_run_script

Run a Kotlin script

kotlin_run_test

Run Kotlin tests for a UniFFI test fixture

python_run_script

Run a Python script

python_run_test

Run Python tests for a UniFFI test fixture

ruby_run_test

Run Ruby tests for a UniFFI test fixture

swift_run_script

Run a Swift script

swift_run_test

Run Swift tests for a UniFFI test fixture