Module uniffi_bindgen::bindings::backend

Backend traits

This module provides a number of traits useful for implementing a backend for Uniffi.

A CodeType is needed for each type that will cross the FFI. It should provide helper machinery in the target language to lift from and lower into a value of that type into a primitive type (the FFIType), and foreign language expressions that call into the machinery. This helper code can be provided by a template file.

A CodeDeclaration is needed for each type that is declared in the UDL file. This has access to the ComponentInterface, which is the closest thing to an Intermediate Representation.

CodeDeclarations provide the target language’s version of the UDL type, including forwarding method calls into Rust. It is likely if you’re implementing a CodeDeclaration for this purpose, it will also need to cross the FFI, and you’ll also need a CodeType.

CodeDeclarations can also be used to conditionally include code: e.g. only include the CallbackInterfaceRuntime if the user has used at least one callback interface.

Each backend has a wrapper template for each file it needs to generate. This should collect the CodeDeclarations that the backend and ComponentInterface between them specify and use them to stitch together a file in the target language.

The CodeOracle provides methods to map the Type values found in the ComponentInterface to the CodeTypes specified by the backend. It also provides methods for transforming identifiers into the coding standard for the target language.

Each backend will have its own filter module, which is used by the askama templates used in all CodeTypes and CodeDeclarations. This filter provides methods to generate expressions and identifiers in the target language. These are all forwarded to the oracle.

Traits

CodeDeclaration
A trait that is able to render a declaration about a particular member declared in the ComponentInterface. Like CodeType, it can render declaration code and imports. It also is able to render code at start-up of the FFI. All methods are optional, and there is no requirement that the trait be used for a particular interface:: member. Thus, it can also be useful for conditionally rendering code.

CodeOracle
An object to look up a foreign language code specific renderer for a given type used. Every Type referred to in the ComponentInterface should map to a corresponding CodeType.

CodeType
A Trait to emit foreign language code to handle referenced types. A type which is specified in the UDL (i.e. a member of the component interface) will have a CodeDeclaration as well, but for types used e.g. primitive types, Strings, etc only a CodeType is needed.

Type Definitions

Literal

TypeIdentifier