Module uniffi_bindgen::bindings::gecko_js::gen_gecko_js

Modules

filters
Filters for our Askama templates above. These output C++ and WebIDL.

Structs

Config
Config options for the generated Firefox front-end bindings. Note that this can only be used to control details that do not affect the underlying component, since the details of the underlying component are entirely determined by the ComponentInterface.

Context
A context associates config options with a component interface, and provides helper methods that are shared between all templates and filters in this module.

Interface
An implementation file generated for each interface in the UDL.

InterfaceHeader
A header file generated for each interface in the UDL.

Namespace
An implementation file for a namespace with top-level functions. If the namespace in the UDL is empty, this isn’t generated.

NamespaceHeader
A header file generated for a namespace containing top-level functions. If the namespace in the UDL file is empty, this file isn’t generated.

SharedHeader
A shared header file that’s included by all our bindings. This defines common serialization logic and extern declarations for the FFI. These namespace and interface source files #include this file.

WebIDL
A template for a Firefox WebIDL file. We only generate one of these per component.