Module uniffi_bindgen::bindings::gecko_js::webidl

This file contains Gecko JS backend-specific extensions for the UniFFI component interface types. These extensions are used to generate the WebIDL file and C++ implementation for Firefox.

Although UDL (the UniFFI Definition Language) and WebIDL share the same syntax, they have completely different semantics. WebIDL distinguishes between “nullable” (T?) and “optional” types (optional T), dictionary members are optional by default, dictionaries are not nullable (but can be optional, and must have a default value if they are). The UniFFI type system is much simpler, but can’t represent these semantics exactly.

The Firefox C++ bindings are also peculiar. A C++ function that implements a WebIDL static or namespace method takes an extra GlobalObject argument; methods that return an ArrayBuffer also take a JSContext; some return values are passed via out parameters while others are returned directly; some WebIDL types map to different C++ types depending on where they’re used (in parameters, out parameters, or dictionary members); and throwing functions take an extra ErrorResult argument.

https://developer.mozilla.org/en-US/docs/Mozilla/WebIDL_bindings describes how Firefox WebIDL constructs are reflected in C++. Since UniFFI is generating both the WebIDL file and the C++ implementation, it must do the same.

These combinations of special cases are impossible to express in Askama, so we have a “shadow type system” with extension traits implemented for the UniFFI interface types. Capturing this logic here lets us keep our templates and filters clean.

Enums

CPPArgument
Describes a function argument.

ReturnBy
Describes how a function returns its result.

ThrowBy
Describes how a function throws errors.

WebIDLType
WebIDL types correspond to UniFFI interface types, but carry additional information for compound types.

Traits

ArgumentExt
Extensions to support WebIDL static method, constructor, and interface method arguments.

ConstructorExt
Extensions to support WebIDL interface constructors.

FieldExt
Extensions to support WebIDL dictionary members.

FunctionExt
Extensions to support WebIDL namespace methods.

MethodExt
Extensions to support WebIDL interface methods.