Trait uniffi_bindgen::bindings::gecko_js::webidl::ArgumentExt

pub trait ArgumentExt {
    fn webidl_type(&self) -> WebIDLType;

    fn optional(&self) -> bool;

    fn webidl_default_value(&self) -> Option<Literal>;
}

Extensions to support WebIDL static method, constructor, and interface method arguments.

Required methods

fn webidl_type(&self) -> WebIDLType

Returns the argument type.

fn optional(&self) -> bool

Indicates if the argument should have an optional keyword. true for nullable dictionaries and arguments that declare a default value in UDL; false otherwise.

fn webidl_default_value(&self) -> Option<Literal>

Returns the default value for this argument, if one is specified.

Loading content...