Trait uniffi_bindgen::bindings::gecko_js::webidl::ConstructorExt

pub trait ConstructorExt {
    fn cpp_arguments(&self) -> Vec<CPPArgument<'_>>;

    fn cpp_throw_by(&self) -> ThrowBy;
}

Extensions to support WebIDL interface constructors.

Required methods

fn cpp_arguments(&self) -> Vec<CPPArgument<'_>>

Returns a list of arguments to declare for this constructor in the C++ implementation, including any extras and out parameters.

fn cpp_throw_by(&self) -> ThrowBy

Indicates how this constructor throws errors, either by an ErrorResult parameter, or by a fatal assertion.

Loading content...