Trait uniffi_bindgen::bindings::gecko_js::webidl::MethodExt

pub trait MethodExt {
    fn webidl_return_type(&self) -> Option<WebIDLType>;

    fn cpp_arguments(&self) -> Vec<CPPArgument<'_>>;

    fn cpp_return_type(&self) -> Option<WebIDLType>;

    fn cpp_return_by(&self) -> ReturnBy;

    fn cpp_throw_by(&self) -> ThrowBy;
}

Extensions to support WebIDL interface methods.

Required methods

fn webidl_return_type(&self) -> Option<WebIDLType>

Returns the WebIDL return type of this method.

fn cpp_arguments(&self) -> Vec<CPPArgument<'_>>

Returns a list of arguments to declare for this method in the C++ implementation, including any extras and out parameters.

fn cpp_return_type(&self) -> Option<WebIDLType>

Returns the C++ return type of this function, or None if the method doesn’t return a value, or returns it via an out parameter.

fn cpp_return_by(&self) -> ReturnBy

Indicates how this function returns its result, either by value or via an out parameter.

fn cpp_throw_by(&self) -> ThrowBy

Indicates how this method throws errors, either by an ErrorResult parameter, or by a fatal assertion.

Loading content...