Module uniffi_bindgen::bindings::kotlin::gen_kotlin::filters

Functions

allocation_size_fn

canonical_name

class_name

Get the idiomatic Kotlin rendering of a class name (for enums, records, errors, etc).

enum_variant

Get the idiomatic Kotlin rendering of an individual enum variant.

exception_name

Get the idiomatic Kotlin rendering of an exception name, replacing Error with Exception.

ffi_converter_name

ffi_type_name

Get the Kotlin syntax for representing a given low-level FfiType.

fn_name

Get the idiomatic Kotlin rendering of a function name.

lift_fn

lower_fn

oracle 🔒

read_fn

render_literal

type_name

unquote

Remove the “`” chars we put around function/variable names

var_name

Get the idiomatic Kotlin rendering of a variable name.

write_fn