Struct uniffi_bindgen::bindings::python::gen_python::TypeRenderer

pub struct TypeRenderer<'a> {
    python_config: &'a Config,
    ci: &'a ComponentInterface,
    include_once_names: RefCell<HashSet<String>>,
    imports: RefCell<BTreeSet<ImportRequirement>>,
}

Renders Python helper code for all types

This template is a bit different than others in that it stores internal state from the render process. Make sure to only call render() once.

Fields
python_config: &'a Configci: &'a ComponentInterfaceinclude_once_names: RefCell<HashSet<String>>imports: RefCell<BTreeSet<ImportRequirement>>
Implementations

impl<'a> TypeRenderer<'a>

fn new(python_config: &'a Config, ci: &'a ComponentInterface) -> Self

fn include_once_check(&self, name: &str) -> bool

fn add_import(&self, name: &str) -> &str

fn add_import_of(&self, mod_name: &str, name: &str) -> &str

fn add_import_of_as( &self, mod_name: &str, symbol_name: &str, as_name: &str ) -> &str

fn get_custom_type_aliases(&self) -> Vec<(String, &Type)>