Module uniffi_bindgen::bindings::swift::gen_swift

Modules

callback_interface 🔒

compounds 🔒

custom 🔒

enum_ 🔒

external 🔒

filters

miscellany 🔒

object 🔒

primitives 🔒

record 🔒

Structs

BridgingHeader

Template for generating the .h file that defines the low-level C FFI.

Config

Config options for the caller to customize the generated Swift.

CustomTypeConfig

ModuleMap

Template for generating the .modulemap file that exposes the low-level C FFI.

SwiftCodeOracle

SwiftWrapper

TypeRenderer

Renders Swift helper code for all types

Statics

ARG_KEYWORDS 🔒

Per https://docs.swift.org/swift-book/documentation/the-swift-programming-language/lexicalstructure/#Keywords-and-Punctuation subset of keywords which need quoting in arg context.

KEYWORDS 🔒

From https://docs.swift.org/swift-book/documentation/the-swift-programming-language/lexicalstructure/#Keywords-and-Punctuation

Traits

CodeType 🔒

A trait tor the implementation.

Functions

generate_bindings

Generate UniFFI component bindings for Swift, as strings in memory.

quote_arg_keyword

Quote a name for use in arg context where fewer keywords must be quoted

quote_general_keyword

Quote a name for use in a context where keywords must be quoted