Struct uniffi_bindgen::bindings::swift::gen_swift::Config

pub struct Config {
    cdylib_name: Option<String>,
    module_name: Option<String>,
    ffi_module_name: Option<String>,
    ffi_module_filename: Option<String>,
    generate_module_map: Option<bool>,
    omit_argument_labels: Option<bool>,
    custom_types: HashMap<String, CustomTypeConfig>,
}

Config options for the caller to customize the generated Swift.

Note that this can only be used to control details of the Swift that do not affect the underlying component, since the details of the underlying component are entirely determined by the ComponentInterface.

Fields
cdylib_name: Option<String>module_name: Option<String>ffi_module_name: Option<String>ffi_module_filename: Option<String>generate_module_map: Option<bool>omit_argument_labels: Option<bool>custom_types: HashMap<String, CustomTypeConfig>
Implementations

impl Config

pub fn module_name(&self) -> String

The name of the Swift module containing the high-level foreign-language bindings.

pub fn ffi_module_name(&self) -> String

The name of the lower-level C module containing the FFI declarations.

pub fn ffi_module_filename(&self) -> String

The filename stem for the lower-level C module containing the FFI declarations.

pub fn modulemap_filename(&self) -> String

The name of the .modulemap file for the lower-level C module with FFI declarations.

pub fn header_filename(&self) -> String

The name of the .h file for the lower-level C module with FFI declarations.

pub fn cdylib_name(&self) -> String

The name of the compiled Rust library containing the FFI implementation.

pub fn generate_module_map(&self) -> bool

Whether to generate a .modulemap file for the lower-level C module with FFI declarations.

pub fn omit_argument_labels(&self) -> bool

Whether to omit argument labels in Swift function definitions.