Struct uniffi_bindgen::bindings::swift::gen_swift::ModuleMap

pub struct ModuleMap<'config, 'ci> {
    config: &'config Config,
    _ci: &'ci ComponentInterface,
}

Template for generating the .modulemap file that exposes the low-level C FFI.

This file defines how the low-level C FFI from BridgingHeader gets exposed as a Swift module that can be called by other Swift code. In our case, its only job is to define the name of the Swift module that will contain the FFI functions so that it can be imported by the higher-level code in from SwiftWrapper.

Fields
config: &'config Config_ci: &'ci ComponentInterface
Implementations

impl<'config, 'ci> ModuleMap<'config, 'ci>

pub fn new(config: &'config Config, _ci: &'ci ComponentInterface) -> Self