Enum uniffi_bindgen::interface::attributes::Attribute

pub(super) enum Attribute {
    ByRef,
    Enum,
    Error,
    Name(String),
    SelfType(SelfType),
    Throws(String),
    External {
        crate_name: String,
        kind: ExternalKind,
    },
    Custom,
    Trait,
}

Represents an attribute parsed from UDL, like [ByRef] or [Throws].

This is a convenience enum for parsing UDL attributes and erroring out if we encounter any unsupported ones. These don’t convert directly into parts of a ComponentInterface, but may influence the properties of things like functions and arguments.

Variants

ByRef

Enum

Error

Name(String)

SelfType(SelfType)

Throws(String)

External

Fields

crate_name: String

kind: ExternalKind

Custom

Trait

Implementations

impl Attribute

pub fn is_error(&self) -> bool

pub fn is_enum(&self) -> bool