Module uniffi_bindgen::interface::callbacks

Callback Interface definitions for a ComponentInterface.

This module converts callback interface definitions from UDL into structures that can be added to a ComponentInterface. A declaration in the UDL like this:

callback interface Example {
  string hello();
};

Will result in a CallbackInterface member being added to the resulting crate::ComponentInterface:

let callback = ci.get_callback_interface_definition("Example").unwrap();
assert_eq!(callback.name(), "Example");
assert_eq!(callback.methods()[0].name(), "hello");

Structs

CallbackInterface

Functions

ffi_callbacks

FfiCallbackFunction functions for the methods of a callback/trait interface

ffi_foreign_future_complete

Definition for callback functions to complete an async callback interface method

foreign_future_ffi_result_struct

Result struct to pass to the completion callback for async methods

method_ffi_callback

method_ffi_callback_name

vtable_name

vtable_struct

FfiStruct for a callback/trait interface VTable