Module uniffi_bindgen::interface::error

Error definitions for a ComponentInterface.

This module converts error definition from UDL into structures that can be added to a ComponentInterface. A declaration in the UDL like this:

[Error]
enum Example {
  "one",
  "two"
};

Will result in an Error member with fieldless variants being added to the resulting ComponentInterface:

let err = ci.get_error_definition("Example").unwrap();
assert_eq!(err.name(), "Example");
assert_eq!(err.variants().len(), 2);
assert_eq!(err.variants()[0].name(), "one");
assert_eq!(err.variants()[1].name(), "two");
assert_eq!(err.is_flat(), true);

A declaration in the UDL like this:

[Error]
interface Example {
  one(i16 code);
  two(string reason);
  three(i32 x, i32 y);
};

Will result in an Error member with variants that have fields being added to the resulting ComponentInterface:

let err = ci.get_error_definition("Example").unwrap();
assert_eq!(err.name(), "Example");
assert_eq!(err.variants().len(), 3);
assert_eq!(err.variants()[0].name(), "one");
assert_eq!(err.variants()[1].name(), "two");
assert_eq!(err.variants()[2].name(), "three");
assert_eq!(err.variants()[0].fields().len(), 0);
assert_eq!(err.variants()[1].fields().len(), 1);
assert_eq!(err.variants()[1].fields()[0].name(), "reason");
assert_eq!(err.variants()[2].fields().len(), 2);
assert_eq!(err.variants()[2].fields()[0].name(), "x");
assert_eq!(err.variants()[2].fields()[1].name(), "y");
assert_eq!(err.is_flat(), false);

Structs

Error

Represents an Error that might be thrown by functions/methods in the component interface.