Enum uniffi_bindgen::interface::ffi::FFIType

pub enum FFIType {

    UInt8,
    Int8,
    UInt16,
    Int16,
    UInt32,
    Int32,
    UInt64,
    Int64,
    Float32,
    Float64,
    RustArcPtr(String),
    RustBuffer,
    ForeignBytes,
    ForeignCallback,

}

Represents the restricted set of low-level types that can be used to construct the C-style FFI layer between a rust component and its foreign language bindings.

For the types that involve memory allocation, we make a distinction between “owned” types (the recipient must free it, or pass it to someone else) and “borrowed” types (the sender must keep it alive for the duration of the call).

Variants

UInt8

Int8

UInt16

Int16

UInt32

Int32

UInt64

Int64

Float32

Float64

RustArcPtr(String)

A *const c_void pointer to a rust-owned Arc<T>. If you’ve got one of these, you must call the appropriate rust function to free it. The templates will generate a unique free function for each T. The inner string references the name of the T type.

RustBuffer

A byte buffer allocated by rust, and owned by whoever currently holds it. If you’ve got one of these, you must either call the appropriate rust function to free it or pass it to someone that will.

ForeignBytes

A borrowed reference to some raw bytes owned by foreign language code. The provider of this reference must keep it alive for the duration of the receiving call.

ForeignCallback

A pointer to a single function in to the foreign language. This function contains all the machinery to make callbacks work on the foreign language side.