Enum uniffi_bindgen::interface::ffi::FfiType

pub enum FfiType {

    UInt8,
    Int8,
    UInt16,
    Int16,
    UInt32,
    Int32,
    UInt64,
    Int64,
    Float32,
    Float64,
    RustArcPtr(String),
    RustBuffer(Option<String>),
    ForeignBytes,
    ForeignCallback,
    ForeignExecutorHandle,
    ForeignExecutorCallback,
    FutureCallback {
        return_type: Box<FfiType>,
    },
    FutureCallbackData,

}

Represents the restricted set of low-level types that can be used to construct the C-style FFI layer between a rust component and its foreign language bindings.

For the types that involve memory allocation, we make a distinction between “owned” types (the recipient must free it, or pass it to someone else) and “borrowed” types (the sender must keep it alive for the duration of the call).

Variants

UInt8

Int8

UInt16

Int16

UInt32

Int32

UInt64

Int64

Float32

Float64

RustArcPtr(String)

A *const c_void pointer to a rust-owned Arc<T>. If you’ve got one of these, you must call the appropriate rust function to free it. The templates will generate a unique free function for each T. The inner string references the name of the T type.

RustBuffer(Option<String>)

A byte buffer allocated by rust, and owned by whoever currently holds it. If you’ve got one of these, you must either call the appropriate rust function to free it or pass it to someone that will. If the inner option is Some, it is the name of the external type. The bindings may need to use this name to import the correct RustBuffer for that type.

ForeignBytes

A borrowed reference to some raw bytes owned by foreign language code. The provider of this reference must keep it alive for the duration of the receiving call.

ForeignCallback

Pointer to a callback function that handles all callbacks on the foreign language side.

ForeignExecutorHandle

Pointer-sized opaque handle that represents a foreign executor. Foreign bindings can either use an actual pointer or a usized integer.

ForeignExecutorCallback

Pointer to the callback function that’s invoked to schedule calls with a ForeignExecutor

FutureCallback

Fields

return_type: Box<FfiType>

Note: return_type is not optional because we have a void callback parameter like we can have a void return. Instead, we use UInt8 as a placeholder value.

Pointer to a callback function to complete an async Rust function

FutureCallbackData

Opaque pointer passed to the FutureCallback