Struct uniffi_bindgen::interface::function::Function

pub struct Function {
    pub(super) name: String,
    pub(super) module_path: String,
    pub(super) is_async: bool,
    pub(super) arguments: Vec<Argument>,
    pub(super) return_type: Option<Type>,
    pub(super) ffi_func: FfiFunction,
    pub(super) throws: Option<Type>,
    pub(super) checksum_fn_name: String,
    pub(super) checksum: Option<u16>,
}

Represents a standalone function.

Each Function corresponds to a standalone function in the rust module, and has a corresponding standalone function in the foreign language bindings.

In the FFI, this will be a standalone function with appropriately lowered types.

Fields
name: Stringmodule_path: Stringis_async: boolarguments: Vec<Argument>return_type: Option<Type>ffi_func: FfiFunctionthrows: Option<Type>checksum_fn_name: Stringchecksum: Option<u16>
Implementations

impl Function

pub fn name(&self) -> &str

pub fn is_async(&self) -> bool

pub fn arguments(&self) -> Vec<&Argument>

pub fn full_arguments(&self) -> Vec<Argument>

pub fn return_type(&self) -> Option<&Type>

pub fn ffi_func(&self) -> &FfiFunction

pub fn checksum_fn_name(&self) -> &str

pub fn checksum(&self) -> u16

pub fn throws(&self) -> bool

pub fn throws_name(&self) -> Option<&str>

pub fn throws_type(&self) -> Option<&Type>

pub fn derive_ffi_func(&mut self) -> Result<()>

pub fn iter_types(&self) -> TypeIterator<'_>