Struct uniffi_bindgen::interface::Enum

pub struct Enum {
    pub(super) name: String,
    pub(super) module_path: String,
    pub(super) discr_type: Option<Type>,
    pub(super) variants: Vec<Variant>,
    pub(super) shape: EnumShape,
    pub(super) non_exhaustive: bool,
    pub(super) docstring: Option<String>,
}

Represents an enum with named variants, each of which may have named and typed fields.

Enums are passed across the FFI by serializing to a bytebuffer, with a i32 indicating the variant followed by the serialization of each field.

Fields
name: Stringmodule_path: Stringdiscr_type: Option<Type>variants: Vec<Variant>shape: EnumShapenon_exhaustive: booldocstring: Option<String>
Implementations

impl Enum

pub fn name(&self) -> &str

pub fn variants(&self) -> &[Variant]

pub fn variant_discr(&self, variant_index: usize) -> Result<Literal>

pub fn variant_discr_type(&self) -> &Option<Type>

pub fn is_flat(&self) -> bool

pub fn is_non_exhaustive(&self) -> bool

pub fn iter_types(&self) -> TypeIterator<'_>

pub fn docstring(&self) -> Option<&str>