Struct uniffi_bindgen::interface::Object

pub struct Object { /* private fields */ }

An “object” is an opaque type that can be instantiated and passed around by reference, have methods called on it, and so on - basically your classic Object Oriented Programming type of deal, except without elaborate inheritence hierarchies.

In UDL these correspond to the interface keyword.

At the FFI layer, objects are represented by an opaque integer handle and a set of functions a common prefix. The object’s constuctors are functions that return new objects by handle, and its methods are functions that take a handle as first argument. The foreign language binding code is expected to stitch these functions back together into an appropriate class definition (or that language’s equivalent thereof).

TODO:

maybe “Class” would be a better name than “Object” here?

Implementations

impl Object

pub fn name(&self) -> &str

pub fn type_(&self) -> Type

pub fn constructors(&self) -> Vec<&Constructor>

pub fn primary_constructor(&self) -> Option<&Constructor>

pub fn alternate_constructors(&self) -> Vec<&Constructor>

pub fn methods(&self) -> Vec<&Method>

pub fn get_method(&self, name: &str) -> Method

pub fn ffi_object_free(&self) -> &FFIFunction

pub fn uses_deprecated_threadsafe_attribute(&self) -> bool

pub fn iter_ffi_function_definitions(
&self
) -> impl Iterator<Item = &FFIFunction>

pub fn derive_ffi_funcs(&mut self, ci_prefix: &str) -> Result<()>

pub fn iter_types(&self) -> TypeIterator<'_>