Struct uniffi_bindgen::interface::RecursiveTypeIterator

struct RecursiveTypeIterator<'a> {
    ci: &'a ComponentInterface,
    current: TypeIterator<'a>,
    seen: HashSet<&'a str>,
    pending: Vec<&'a Type>,
}

Stateful iterator for yielding all types contained in a given item.

This struct is the implementation of ComponentInterface::iter_types_in_item and should be considered an opaque implementation detail. It’s a separate struct because I couldn’t figure out a way to implement it using iterators and closures that would make the lifetimes work out correctly.

The idea here is that we want to yield all the types from IterTypes::iter_types on a given item, and additionally we want to recurse into the definition of any user-provided types like records, enums, etc so we can also yield the types contained therein.

To guard against infinite recursion, we maintain a list of previously-seen user-defined types, ensuring that we recurse into the definition of those types only once. To simplify the implementation, we maintain a queue of pending user-defined types that we have seen but not yet recursed into. (Ironically, the use of an explicit queue means our implementation is not actually recursive…)

Fields
ci: &'a ComponentInterface

The ComponentInterface from which this iterator was created.

current: TypeIterator<'a>

The currently-active iterator from which we’re yielding.

seen: HashSet<&'a str>

A set of names of user-defined types that we have already seen.

pending: Vec<&'a Type>

A queue of user-defined types that we need to recurse into.

Implementations

impl<'a> RecursiveTypeIterator<'a>

fn new<T: IterTypes>(
ci: &'a ComponentInterface,
item: &'a T
) -> RecursiveTypeIterator<'a>

Allocate a new RecursiveTypeIterator over the given item.

fn add_pending_type(&mut self, type_: &'a Type)

Add a new type to the queue of pending types, if not previously seen.

fn advance_to_next_type(&mut self) -> Option<&'a Type>

Advance the iterator to recurse into the next pending type, if any.

This method is called when the current iterator is empty, and it will select the next pending type from the queue and start iterating over its contained types. The return value will be the first item from the new iterator.