Enum uniffi_bindgen::interface::types::Type

pub enum Type {

    UInt8,
    Int8,
    UInt16,
    Int16,
    UInt32,
    Int32,
    UInt64,
    Int64,
    Float32,
    Float64,
    Boolean,
    String,
    Bytes,
    Timestamp,
    Duration,
    Object {
        name: String,
        imp: ObjectImpl,
    },
    ForeignExecutor,
    Record(String),
    Enum(String),
    Error(String),
    CallbackInterface(String),
    Optional(Box<Type>),
    Sequence(Box<Type>),
    Map(Box<Type>, Box<Type>),
    External {
        name: String,
        crate_name: String,
        kind: ExternalKind,
    },
    Custom {
        name: String,
        builtin: Box<Type>,
    },

}

Represents all the different high-level types that can be used in a component interface. At this level we identify user-defined types by name, without knowing any details of their internal structure apart from what type of thing they are (record, enum, etc).

Variants

UInt8

Int8

UInt16

Int16

UInt32

Int32

UInt64

Int64

Float32

Float64

Boolean

String

Bytes

Timestamp

Duration

Object

Fields

name: String

imp: ObjectImpl

ForeignExecutor

Record(String)

Enum(String)

Error(String)

CallbackInterface(String)

Optional(Box<Type>)

Sequence(Box<Type>)

Map(Box<Type>, Box<Type>)

External

Fields

name: String

crate_name: String

kind: ExternalKind

Custom

Fields

name: String

builtin: Box<Type>

Implementations

impl Type

pub fn canonical_name(&self) -> String

Get the canonical, unique-within-this-component name for a type.

When generating helper code for foreign language bindings, it’s sometimes useful to be able to name a particular type in order to e.g. call a helper function that is specific to that type. We support this by defining a naming convention where each type gets a unique canonical name, constructed recursively from the names of its component types (if any).

pub fn ffi_type(&self) -> FfiType

pub fn iter_types(&self) -> TypeIterator<'_>