Module uniffi_bindgen::interface::types::finder

Helpers for finding the named types defined in a UDL interface.

This module provides the TypeFinder trait, an abstraction for walking the weedle parse tree, looking for type definitions, and accumulating them in a TypeUniverse.

The type-finding process only discovers very basic information about names and their corresponding types. For example, it can discover that “Foobar” names a Record, but it won’t discover anything about the fields of that record.

Factoring this functionality out into a separate phase makes the subsequent work of more detailed parsing of the UDL a lot simpler, we know how to resolve names to types when building up the full interface definition.

Traits

TypeFinder

Trait to help with an early “type discovery” phase when processing the UDL.