Type Alias uniffi_bindgen::interface::universe::TypeIterator

pub type TypeIterator<'a> = Box<dyn Iterator<Item = &'a Type> + 'a>;

An abstract type for an iterator over &Type references.

Ideally we would not need to name this type explicitly, and could just use an impl Iterator<Item = &Type> on any method that yields types.

Aliased Type

struct TypeIterator<'a>(Unique<dyn Iterator<Item = &'a Type> + 'a>, Global);

Fields
0: Unique<dyn Iterator<Item = &'a Type> + 'a>1: Global