Trait uniffi_bindgen::BindingsConfig

pub trait BindingsConfig: DeserializeOwned {
    const TOML_KEY: &'static str;

    fn update_from_ci(&mut self, ci: &ComponentInterface);
    fn update_from_cdylib_name(&mut self, cdylib_name: &str);
    fn update_from_dependency_configs(
        &mut self,
        config_map: HashMap<&str, &Self>
    );
}

Trait for bindings configuration. Each bindings language defines one of these.

BindingsConfigs are initially loaded from uniffi.toml file. Then the trait methods are used to fill in missing values.

Required Associated Constants

const TOML_KEY: &'static str

key in the bindings table from uniffi.toml for this configuration

Required Methods

fn update_from_ci(&mut self, ci: &ComponentInterface)

Update missing values using the ComponentInterface

fn update_from_cdylib_name(&mut self, cdylib_name: &str)

Update missing values using the dylib file for the main crate, when in library mode.

cdylib_name will be the library filename without the leading lib and trailing extension

fn update_from_dependency_configs(&mut self, config_map: HashMap<&str, &Self>)

Update missing values from config instances from dependent crates

config_map maps crate names to config instances. This is mostly used to set up external types.