Trait uniffi_bindgen::backend::CodeOracle

pub trait CodeOracle {
    // Required methods
    fn find(&self, type_: &Type) -> Box<dyn CodeType>;
    fn class_name(&self, nm: &str) -> String;
    fn fn_name(&self, nm: &str) -> String;
    fn var_name(&self, nm: &str) -> String;
    fn enum_variant_name(&self, nm: &str) -> String;
    fn error_name(&self, nm: &str) -> String;
    fn ffi_type_label(&self, ffi_type: &FfiType) -> String;
}

An object to supply a foreign language specific CodeType for a given type. It also supplys the specific rendering of a given identifier when used in a specific context.

Required Methods

fn find(&self, type_: &Type) -> Box<dyn CodeType>

fn class_name(&self, nm: &str) -> String

Get the idiomatic rendering of a class name (for enums, records, errors, etc).

fn fn_name(&self, nm: &str) -> String

Get the idiomatic rendering of a function name.

fn var_name(&self, nm: &str) -> String

Get the idiomatic rendering of a variable name.

fn enum_variant_name(&self, nm: &str) -> String

Get the idiomatic rendering of an individual enum variant.

fn error_name(&self, nm: &str) -> String

Get the idiomatic rendering of an error name.

fn ffi_type_label(&self, ffi_type: &FfiType) -> String