Trait uniffi_bindgen::backend::CodeOracle

pub trait CodeOracle {
    fn find(&self, type_: &TypeIdentifier) -> Box<dyn CodeType>;

    fn class_name(&self, nm: &str) -> String;

    fn fn_name(&self, nm: &str) -> String;

    fn var_name(&self, nm: &str) -> String;

    fn enum_variant_name(&self, nm: &str) -> String;

    fn error_name(&self, nm: &str) -> String;

    fn ffi_type_label(&self, ffi_type: &FFIType) -> String;
}

An object to look up a foreign language code specific renderer for a given type used. Every Type referred to in the ComponentInterface should map to a corresponding CodeType.

The mapping may be opaque, but the oracle always knows the answer.

In adddition, the oracle knows how to render identifiers (function names, class names, variable names etc).

Required methods

fn find(&self, type_: &TypeIdentifier) -> Box<dyn CodeType>

fn class_name(&self, nm: &str) -> String

Get the idiomatic rendering of a class name (for enums, records, errors, etc).

fn fn_name(&self, nm: &str) -> String

Get the idiomatic rendering of a function name.

fn var_name(&self, nm: &str) -> String

Get the idiomatic rendering of a variable name.

fn enum_variant_name(&self, nm: &str) -> String

Get the idiomatic rendering of an individual enum variant.

fn error_name(&self, nm: &str) -> String

Get the idiomatic rendering of an error name.

fn ffi_type_label(&self, ffi_type: &FFIType) -> String