Trait uniffi_bindgen::backend::CodeType

pub trait CodeType {
    fn type_label(&self, oracle: &dyn CodeOracle) -> String;

    fn canonical_name(&self, oracle: &dyn CodeOracle) -> String { ... }
    fn literal(&self, oracle: &dyn CodeOracle, _literal: &Literal) -> String { ... }
    fn ffi_converter_name(&self, oracle: &dyn CodeOracle) -> String { ... }
    fn lower(&self, oracle: &dyn CodeOracle, _nm: &str) -> String { ... }
    fn write(&self, oracle: &dyn CodeOracle, _nm: &str, _target: &str) -> String { ... }
    fn lift(&self, oracle: &dyn CodeOracle, _nm: &str) -> String { ... }
    fn read(&self, oracle: &dyn CodeOracle, _nm: &str) -> String { ... }
    fn helper_code(&self, _oracle: &dyn CodeOracle) -> Option<String> { ... }
    fn imports(&self, _oracle: &dyn CodeOracle) -> Option<Vec<String>> { ... }
    fn initialization_fn(&self, _oracle: &dyn CodeOracle) -> Option<String> { ... }
    fn coerce(&self, oracle: &dyn CodeOracle, _nm: &str) -> String { ... }
}

A Trait to emit foreign language code to handle referenced types. A type which is specified in the UDL (i.e. a member of the component interface) will have a CodeDeclaration as well, but for types used e.g. primitive types, Strings, etc only a CodeType is needed.

Required Methods

fn type_label(&self, oracle: &dyn CodeOracle) -> String

The language specific label used to reference this type. This will be used in method signatures and property declarations.

Provided Methods

fn canonical_name(&self, oracle: &dyn CodeOracle) -> String

A representation of this type label that can be used as part of another identifier. e.g. read_foo(), or FooInternals.

This is especially useful when creating specialized objects or methods to deal with this type only.

fn literal(&self, oracle: &dyn CodeOracle, _literal: &Literal) -> String

A representation of the given literal for this type. N.B. Literal is aliased from interface::Literal, so may not be whole suited to this task.

fn ffi_converter_name(&self, oracle: &dyn CodeOracle) -> String

Name of the FfiConverter

This is the object that contains the lower, write, lift, and read methods for this type. Depending on the binding this will either be a singleton or a class with static methods.

This is the newer way of handling these methods and replaces the lower, write, lift, and read CodeType methods. Currently only used by Kotlin, but the plan is to move other backends to using this.

fn lower(&self, oracle: &dyn CodeOracle, _nm: &str) -> String

An expression for lowering a value into something we can pass over the FFI.

N.B. This should align with the helper_code generated by this CodeType.

fn write(&self, oracle: &dyn CodeOracle, _nm: &str, _target: &str) -> String

An expression for writing a value into a byte buffer.

N.B. This should align with the helper_code generated by this CodeType.

fn lift(&self, oracle: &dyn CodeOracle, _nm: &str) -> String

An expression for lifting a value from something we received over the FFI.

N.B. This should align with the helper_code generated by this CodeType.

fn read(&self, oracle: &dyn CodeOracle, _nm: &str) -> String

An expression for reading a value from a byte buffer.

N.B. This should align with the helper_code generated by this CodeType.

fn helper_code(&self, _oracle: &dyn CodeOracle) -> Option<String>

The lift/lower/read/write methods above must be producing expressions that can be part of a larger statement. Most of the time, that is a function call to do the work for it. The functions being called by those experessions should be declared in the helper code generated by this method.

fn imports(&self, _oracle: &dyn CodeOracle) -> Option<Vec<String>>

A list of imports that are needed if this type is in use. Classes are imported exactly once.

fn initialization_fn(&self, _oracle: &dyn CodeOracle) -> Option<String>

Function to run at startup

fn coerce(&self, oracle: &dyn CodeOracle, _nm: &str) -> String

An expression to coerce the given variable to the expected type.