Module uniffi_bindgen::backend::types

Backend traits

This module provides a number of traits useful for implementing a backend for Uniffi.

A CodeType is needed for each type that will cross the FFI. It should provide helper machinery in the target language to lift from and lower into a value of that type into a primitive type (the FFIType), and foreign language expressions that call into the machinery. This helper code can be provided by a template file.

A CodeDeclaration is needed for each type that is declared in the UDL file. This has access to the crate::interface::ComponentInterface, which is the closest thing to an Intermediate Representation.

CodeDeclarations provide the target language’s version of the UDL type, including forwarding method calls into Rust. It is likely if you’re implementing a CodeDeclaration for this purpose, it will also need to cross the FFI, and you’ll also need a CodeType.

CodeDeclarations can also be used to conditionally include code: e.g. only include the CallbackInterfaceRuntime if the user has used at least one callback interface.

Each backend has a wrapper template for each file it needs to generate. This should collect the CodeDeclarations that the backend and ComponentInterface between them specify and use them to stitch together a file in the target language.

The CodeOracle provides methods to map the Type values found in the ComponentInterface to the CodeTypes specified by the backend. It also provides methods for transforming identifiers into the coding standard for the target language.

Each backend will have its own filter module, which is used by the askama templates used in all CodeTypes and CodeDeclarations. This filter provides methods to generate expressions and identifiers in the target language. These are all forwarded to the oracle.

Traits

CodeType

A Trait to emit foreign language code to handle referenced types. A type which is specified in the UDL (i.e. a member of the component interface) will have a CodeDeclaration as well, but for types used e.g. primitive types, Strings, etc only a CodeType is needed.

CodeTypeDispatch

This trait is used to implement CodeType for Type and type-like structs (Record, Enum, Field, etc). We forward all method calls to a Box<dyn CodeType>, which we get by calling CodeOracle.find().