Trait uniffi_bindgen::bindings::backend::CodeType

pub trait CodeType {
    fn type_label(&self, oracle: &dyn CodeOracle) -> String;

    fn canonical_name(&self, oracle: &dyn CodeOracle) -> String { ... }

    fn literal(&self, oracle: &dyn CodeOracle, _literal: &Literal) -> String { ... }

    fn lower(&self, oracle: &dyn CodeOracle, _nm: &dyn Display) -> String { ... }

    fn write(
        &self, 
        oracle: &dyn CodeOracle, 
        _nm: &dyn Display, 
        _target: &dyn Display
    ) -> String { ... }

    fn lift(&self, oracle: &dyn CodeOracle, _nm: &dyn Display) -> String { ... }

    fn read(&self, oracle: &dyn CodeOracle, _nm: &dyn Display) -> String { ... }

    fn helper_code(&self, _oracle: &dyn CodeOracle) -> Option<String> { ... }

    fn imports(&self, _oracle: &dyn CodeOracle) -> Option<Vec<String>> { ... }
}

A Trait to emit foreign language code to handle referenced types. A type which is specified in the UDL (i.e. a member of the component interface) will have a CodeDeclaration as well, but for types used e.g. primitive types, Strings, etc only a CodeType is needed.

Required methods

fn type_label(&self, oracle: &dyn CodeOracle) -> String

The language specific label used to reference this type. This will be used in method signatures and property declarations.

Loading content...
Provided methods

fn canonical_name(&self, oracle: &dyn CodeOracle) -> String

A representation of this type label that can be used as part of another identifier. e.g. read_foo(), or FooInternals.

This is especially useful when creating specialized objects or methods to deal with this type only.

fn literal(&self, oracle: &dyn CodeOracle, _literal: &Literal) -> String

A representation of the given literal for this type. N.B. Literal is aliased from interface::Literal, so may not be whole suited to this task.

fn lower(&self, oracle: &dyn CodeOracle, _nm: &dyn Display) -> String

An expression for lowering a value into something we can pass over the FFI.

N.B. This should align with the helper_code generated by this CodeType.

fn write(
&self,
oracle: &dyn CodeOracle,
_nm: &dyn Display,
_target: &dyn Display
) -> String

An expression for writing a value into a byte buffer.

N.B. This should align with the helper_code generated by this CodeType.

fn lift(&self, oracle: &dyn CodeOracle, _nm: &dyn Display) -> String

An expression for lifting a value from something we received over the FFI.

N.B. This should align with the helper_code generated by this CodeType.

fn read(&self, oracle: &dyn CodeOracle, _nm: &dyn Display) -> String

An expression for reading a value from a byte buffer.

N.B. This should align with the helper_code generated by this CodeType.

fn helper_code(&self, _oracle: &dyn CodeOracle) -> Option<String>

The lift/lower/read/write methods above must be producing expressions that can be part of a larger statement. Most of the time, that is a function call to do the work for it. The functions being called by those experessions should be declared in the helper code generated by this method.

fn imports(&self, _oracle: &dyn CodeOracle) -> Option<Vec<String>>

A list of imports that are needed if this type is in use. Classes are imported exactly once.

Loading content...