Module uniffi_bindgen::bindings::kotlin::gen_kotlin::filters

Functions

canonical_name

class_name_kt
Get the idiomatic Kotlin rendering of a class name (for enums, records, errors, etc).

enum_variant_kt
Get the idiomatic Kotlin rendering of an individual enum variant.

exception_name_kt
Get the idiomatic Kotlin rendering of an exception name

fn_name_kt
Get the idiomatic Kotlin rendering of a function name.

lift_kt

literal_kt

lower_kt

oracle

read_kt

type_ffi
Get the Kotlin syntax for representing a given low-level FFIType.

type_kt

var_name_kt
Get the idiomatic Kotlin rendering of a variable name.

write_kt