Module uniffi_bindgen::bindings::kotlin::gen_kotlin::filters

Re-exports

pub use crate::backend::filters::*;

Functions

allocation_size_fn

canonical_name

enum_variant

Get a codetype for idiomatic Kotlin rendering of an individual enum variant.

error_canonical_name

error_ffi_converter_name

error_handler

error_type_name

Some of the above filters have different versions to help when the type is used as an error.

error_variant

Get a codetype for idiomatic Kotlin rendering of an individual enum variant when used in an error.

ffi_converter_name

ffi_type_name

Get the Kotlin syntax for representing a given low-level FfiType.

ffi_type_name_by_value

fn_name

Get the idiomatic Kotlin rendering of a function name.

future_callback_handler

future_continuation_type

lift_fn

lower_fn

object_names

read_fn

render_literal

type_name

unique_ffi_types

unquote

Remove the “`” chars we put around function/variable names

var_name

Get the idiomatic Kotlin rendering of a variable name.

variant_name

write_fn