Module uniffi_bindgen::bindings::swift::gen_swift

Modules

callback_interface 🔒

compounds 🔒

custom 🔒

enum_ 🔒

error 🔒

executor 🔒

external 🔒

filters

miscellany 🔒

object 🔒

primitives 🔒

record 🔒

Structs

BridgingHeader

Template for generating the .h file that defines the low-level C FFI.

Config

Config options for the caller to customize the generated Swift.

CustomTypeConfig

ModuleMap

Template for generating the .modulemap file that exposes the low-level C FFI.

SwiftCodeOracle

SwiftWrapper

TypeRenderer

Renders Swift helper code for all types

Functions

generate_bindings

Generate UniFFI component bindings for Swift, as strings in memory.