Module uniffi_bindgen::bindings::swift::gen_swift::filters

Functions

canonical_name

class_name

Get the idiomatic Swift rendering of a class name (for enums, records, errors, etc).

enum_variant_swift

Get the idiomatic Swift rendering of an individual enum variant.

ffi_converter_name

ffi_type_name

Get the Swift syntax for representing a given low-level FfiType.

fn_name

Get the idiomatic Swift rendering of a function name.

lift_fn

literal_swift

lower_fn

oracle 🔒

read_fn

type_ffi_lowered

Get the type that a type is lowered into. This is subtly different than type_ffi, see #1106 for details

type_name

var_name

Get the idiomatic Swift rendering of a variable name.

write_fn