Module uniffi_bindgen::bindings::swift::gen_swift::filters

Functions

canonical_name

class_name_swift
Get the idiomatic Swift rendering of a class name (for enums, records, errors, etc).

enum_variant_swift
Get the idiomatic Swift rendering of an individual enum variant.

exception_name_swift
Get the idiomatic Swift rendering of an exception name

fn_name_swift
Get the idiomatic Swift rendering of a function name.

lift_swift

literal_swift

lower_swift

oracle

read_swift

type_ffi
Get the Swift syntax for representing a given low-level FFIType.

type_swift

var_name_swift
Get the idiomatic Swift rendering of a variable name.

write_swift