Module uniffi_bindgen::interface::attributes

Attribute definitions for a ComponentInterface.

This module provides some conveniences for working with attribute definitions from WebIDL. When encountering a weedle ExtendedAttribute node, use TryFrom to convert it into an Attribute representing one of the attributes that we support. You can also use the parse_attributes function to parse an ExtendedAttributeList into a vec of same.

We only support a small number of attributes, so it’s manageable to have them all handled by a single abstraction. This might need to be refactored in future if we grow significantly more complicated attribute handling.

Structs

ArgumentAttributes 🔒

Represents UDL attributes that might appear on a function argument.

ConstructorAttributes 🔒

Represents UDL attributes that might appear on a constructor.

EnumAttributes 🔒

Attributes that can be attached to an enum definition in the UDL. There’s only one case here: using [Error] to mark an enum as an error class.

FunctionAttributes 🔒

Represents UDL attributes that might appear on a function.

InterfaceAttributes 🔒

Represents UDL attributes that might appear on an interface definition.

MethodAttributes 🔒

Represents UDL attributes that might appear on a method.

TypedefAttributes 🔒

Represents UDL attributes that might appear on a typedef

Enums

Attribute 🔒

Represents an attribute parsed from UDL, like [ByRef] or [Throws].

SelfType 🔒

Represents the different possible types of method call receiver.

Functions

name_from_id_or_string 🔒

parse_attributes 🔒

Parse a weedle ExtendedAttributeList into a list of Attributes, erroring out on duplicates.