Module uniffi_bindgen::interface::enum_

Enum definitions for a ComponentInterface.

This module converts enum definition from UDL into structures that can be added to a ComponentInterface. A declaration in the UDL like this:

enum Example {
  "one",
  "two"
};

Will result in a Enum member being added to the resulting ComponentInterface:

let e = ci.get_enum_definition("Example").unwrap();
assert_eq!(e.name(), "Example");
assert_eq!(e.variants().len(), 2);
assert_eq!(e.variants()[0].name(), "one");
assert_eq!(e.variants()[1].name(), "two");

Like in Rust, UniFFI enums can contain associated data, but this needs to be declared with a different syntax in order to work within the restrictions of WebIDL. A declaration like this:

[Enum]
interface Example {
  Zero();
  One(u32 first);
  Two(u32 first, string second);
};

Will result in an Enum member whose variants have associated fields:

let e = ci.get_enum_definition("ExampleWithData").unwrap();
assert_eq!(e.name(), "ExampleWithData");
assert_eq!(e.variants().len(), 3);
assert_eq!(e.variants()[0].name(), "Zero");
assert_eq!(e.variants()[0].fields().len(), 0);
assert_eq!(e.variants()[1].name(), "One");
assert_eq!(e.variants()[1].fields().len(), 1);
assert_eq!(e.variants()[1].fields()[0].name(), "first");

Structs

Enum
Represents an enum with named variants, each of which may have named and typed fields.

Variant
Represents an individual variant in an Enum.