Module uniffi_bindgen::interface::ffi

Low-level typesystem for the FFI layer of a component interface.

This module provides the “FFI-level” typesystem of a UniFFI Rust Component, that is, the C-style functions and structs and primitive datatypes that are used to interface between the Rust component code and the foreign-language bindings.

These types are purely an implementation detail of UniFFI, so consumers shouldn’t need to know about them. But as a developer working on UniFFI itself, you’re likely to spend a lot of time thinking about how these low-level types are used to represent the higher-level “interface types” from the Type enum.

Structs

FfiArgument

Represents an argument to an FFI function.

FfiCallbackFunction

Represents an “extern C”-style callback function

FfiField

Represents a field of an FfiStruct

FfiFunction

Represents an “extern C”-style function that will be part of the FFI.

FfiStruct

Represents a repr(C) struct used in the FFI

Enums

FfiDefinition

An Ffi definition

FfiType