Module uniffi_bindgen::interface::namespace

Namespace definition for a ComponentInterface.

This module converts a namespace definition from UDL into structures that can be added to a ComponentInterface.

In WebIDL proper, each namespace declares a set of functions and attributes that are exposed as a global object of that name, and there can be any number of such namespace definitions.

For our purposes with UDL, we expect just a single namespace declaration, which defines properties of the component as a whole (currently just the name). It also contains the functions that will be exposed as individual plain functions exported by the component, if any. So something like this:

namespace example {
  string hello();
};

Declares a component named “example” with a single exported function named “hello”:

assert_eq!(ci.namespace(), "example");
assert_eq!(ci.get_function_definition("hello").unwrap().name(), "hello");

While this awkward-looking syntax:

namespace example {};

Declares a component named “example” with no exported functions:

assert_eq!(ci.namespace(), "example");
assert_eq!(ci.iter_function_definitions().len(), 0);

Yeah, it’s a bit of an awkward fit syntactically, but it’s enough to get us up and running for a first version of this tool.

Structs

Namespace
A namespace is currently just a name, but might hold more metadata about the component in future.