Struct uniffi_bindgen::interface::ComponentInterface

pub struct ComponentInterface { /* private fields */ }

The main public interface for this module, representing the complete details of an interface exposed by a rust component and the details of consuming it via an extern-C FFI layer.

Implementations

impl ComponentInterface

pub fn from_webidl(idl: &str) -> Result<Self>

Parse a ComponentInterface from a string containing a WebIDL definition.

pub fn namespace(&self) -> &str

The string namespace within which this API should be presented to the caller.

This string would typically be used to prefix function names in the FFI, to build a package or module name for the foreign language, etc.

pub fn enum_definitions(&self) -> &[Enum]

Get the definitions for every Enum type in the interface.

pub fn get_enum_definition(&self, name: &str) -> Option<&Enum>

Get an Enum definition by name, or None if no such Enum is defined.

pub fn record_definitions(&self) -> &[Record]

Get the definitions for every Record type in the interface.

pub fn get_record_definition(&self, name: &str) -> Option<&Record>

Get a Record definition by name, or None if no such Record is defined.

pub fn function_definitions(&self) -> &[Function]

Get the definitions for every Function in the interface.

pub fn get_function_definition(&self, name: &str) -> Option<&Function>

Get a Function definition by name, or None if no such Function is defined.

pub fn object_definitions(&self) -> &[Object]

Get the definitions for every Object type in the interface.

pub fn get_object_definition(&self, name: &str) -> Option<&Object>

Get an Object definition by name, or None if no such Object is defined.

pub fn callback_interface_definitions(&self) -> &[CallbackInterface]

Get the definitions for every Callback Interface type in the interface.

pub fn get_callback_interface_definition(
&self,
name: &str
) -> Option<&CallbackInterface>

Get a Callback interface definition by name, or None if no such interface is defined.

pub fn error_definitions(&self) -> &[Error]

Get the definitions for every Error type in the interface.

pub fn get_error_definition(&self, name: &str) -> Option<&Error>

Get an Error definition by name, or None if no such Error is defined.

pub fn iter_external_types(&self) -> impl Iterator<Item = (&String, &String)>

Get details about all Type::External types

pub fn iter_custom_types(&self) -> impl Iterator<Item = (&String, &Type)>

Get details about all Type::Custom types

pub fn iter_types(&self) -> impl Iterator<Item = &Type>

Iterate over all known types in the interface.

pub fn get_type(&self, name: &str) -> Option<Type>

Get a specific type

pub fn item_contains_object_references(&self, item: &Type) -> bool

Check whether the given item contains any (possibly nested) Type::Object references.

This is important to know in language bindings that cannot integrate object types tightly with the host GC, and hence need to perform manual destruction of objects.

pub fn item_contains_unsigned_types(&self, item: &Type) -> bool

Check whether the given item contains any (possibly nested) unsigned types

pub fn contains_optional_types(&self) -> bool

Check whether the interface contains any optional types

pub fn contains_sequence_types(&self) -> bool

Check whether the interface contains any sequence types

pub fn contains_map_types(&self) -> bool

Check whether the interface contains any map types

pub fn checksum(&self) -> u64

Calculate a numeric checksum for this ComponentInterface.

The checksum can be used to guard against accidentally using foreign-language bindings generated from one version of an interface with the compiled Rust code from a different version of that interface. It offers the following properties:

Two ComponentIntefaces generated from the same WebIDL file, using the same version of uniffi and the same version of Rust, will always have the same checksum value.

Two ComponentInterfaces will, with high probability, have different checksum values if: - They were generated from two different WebIDL files. - They were generated by two different versions of uniffi

The checksum may or may not change depending on the version of Rust used; since we expect consumers to be using the same executable to generate both the scaffolding and the bindings, assuming the same version of Rust seems acceptable.

Note that this is designed to prevent accidents, not attacks, so there is no need for the checksum to be cryptographically secure.

TODO: it’s not clear to me if the derivation of Hash is actually deterministic enough to ensure the guarantees above, or if it might be sensitive to e.g. compiler-driven re-ordering of struct field. Let’s see how it goes…

pub fn ffi_namespace(&self) -> String

The namespace to use in FFI-level function definitions.

The value returned by this method is used as a prefix to namespace all FFI-level functions used in this ComponentInterface.

Since these names are an internal implementation detail that is not typically visible to consumers, we take the opportunity to add an additional safety guard by including a 4-hex-char checksum in each name. If foreign-language bindings attempt to load and use a version of the Rust code compiled from a different UDL definition than the one used for the bindings themselves, then there is a high probability of checksum mismatch and they will fail to link against the compiled Rust code. The result will be an ugly inscrutable link-time error, but that is a lot better than triggering potentially arbitrary memory unsafety!

pub fn ffi_rustbuffer_alloc(&self) -> FFIFunction

Builtin FFI function for allocating a new RustBuffer. This is needed so that the foreign language bindings can create buffers in which to pass complex data types across the FFI.

pub fn ffi_rustbuffer_from_bytes(&self) -> FFIFunction

Builtin FFI function for copying foreign-owned bytes This is needed so that the foreign language bindings can create buffers in which to pass complex data types across the FFI.

pub fn ffi_rustbuffer_free(&self) -> FFIFunction

Builtin FFI function for freeing a RustBuffer. This is needed so that the foreign language bindings can free buffers in which they received complex data types returned across the FFI.

pub fn ffi_rustbuffer_reserve(&self) -> FFIFunction

Builtin FFI function for reserving extra space in a RustBuffer. This is needed so that the foreign language bindings can grow buffers used for passing complex data types across the FFI.

pub fn iter_ffi_function_definitions(
&self
) -> impl Iterator<Item = FFIFunction> + '_

List the definitions of all FFI functions in the interface.

The set of FFI functions is derived automatically from the set of higher-level types along with the builtin FFI helper functions.

pub fn iter_user_ffi_function_definitions(
&self
) -> impl Iterator<Item = &FFIFunction> + '_

List all FFI functions definitions for user-defined interfaces

This includes FFI functions for:

Top-level functions

Object methods

Callback interfaces

pub fn iter_rust_buffer_ffi_function_definitions(
&self
) -> impl Iterator<Item = FFIFunction>

List all FFI functions definitions for RustBuffer functionality