Struct uniffi_bindgen::interface::ComponentInterface

pub struct ComponentInterface {
    uniffi_version: String,
    types: TypeUniverse,
    namespace: String,
    enums: Vec<Enum>,
    records: Vec<Record>,
    functions: Vec<Function>,
    objects: Vec<Object>,
    callback_interfaces: Vec<CallbackInterface>,
    errors: Vec<Error>,
}

The main public interface for this module, representing the complete details of an interface exposed by a rust component and the details of consuming it via an extern-C FFI layer.

Fields
uniffi_version: String

Every ComponentInterface gets tagged with the version of uniffi used to create it. This helps us avoid using a lib compiled with one version together with bindings created using a different version, which might introduce unsafety.

types: TypeUniverse

All of the types used in the interface.

namespace: String

The unique prefix that we’ll use for namespacing when exposing this component’s API.

enums: Vec<Enum>

The high-level API provided by the component.

records: Vec<Record>functions: Vec<Function>objects: Vec<Object>callback_interfaces: Vec<CallbackInterface>errors: Vec<Error>
Implementations

impl<'ci> ComponentInterface

pub fn from_webidl(idl: &str) -> Result<Self>

Parse a ComponentInterface from a string containing a WebIDL definition.

pub fn namespace(&self) -> &str

The string namespace within which this API should be presented to the caller.

This string would typically be used to prefix function names in the FFI, to build a package or module name for the foreign language, etc.

pub fn iter_enum_definitions(&self) -> Vec<Enum>

List the definitions for every Enum type in the interface.

pub fn get_enum_definition(&self, name: &str) -> Option<&Enum>

Get an Enum definition by name, or None if no such Enum is defined.

pub fn iter_record_definitions(&self) -> Vec<Record>

List the definitions for every Record type in the interface.

pub fn get_record_definition(&self, name: &str) -> Option<&Record>

Get a Record definition by name, or None if no such Record is defined.

pub fn iter_function_definitions(&self) -> Vec<Function>

List the definitions for every Function in the interface.

pub fn get_function_definition(&self, name: &str) -> Option<&Function>

Get a Function definition by name, or None if no such Function is defined.

pub fn iter_object_definitions(&self) -> Vec<Object>

List the definitions for every Object type in the interface.

pub fn get_object_definition(&self, name: &str) -> Option<&Object>

Get an Object definition by name, or None if no such Object is defined.

pub fn iter_callback_interface_definitions(&self) -> Vec<CallbackInterface>

List the definitions for every Callback Interface type in the interface.

pub fn get_callback_interface_definition(
&self,
name: &str
) -> Option<&CallbackInterface>

Get a Callback interface definition by name, or None if no such interface is defined.

pub fn iter_error_definitions(&self) -> Vec<Error>

List the definitions for every Error type in the interface.

pub fn get_error_definition(&self, name: &str) -> Option<&Error>

Get an Error definition by name, or None if no such Error is defined.

pub fn iter_external_types(&self) -> Vec<(String, String)>

Get details about all Type::External types

pub fn iter_custom_types(&self) -> Vec<(String, Type)>

Get details about all Type::Custom types

pub fn iter_types(&self) -> Vec<Type>

Iterate over all known types in the interface.

pub fn get_type(&self, name: &str) -> Option<Type>

Get a specific type

fn iter_types_in_item<'a, T: IterTypes>(
&'a self,
item: &'a T
) -> impl Iterator<Item = &'a Type> + 'a

Iterate over all types contained in the given item.

This method uses IterTypes::iter_types to iterate over the types contained within the given item, but additionally recurses into the definition of user-defined types like records and enums to yield the types that they contain.

pub fn item_contains_object_references<T: IterTypes>(&self, item: &T) -> bool

Check whether the given item contains any (possibly nested) Type::Object references.

This is important to know in language bindings that cannot integrate object types tightly with the host GC, and hence need to perform manual destruction of objects.

pub fn item_contains_unsigned_types<T: IterTypes>(&self, item: &T) -> bool

Check whether the given item contains any (possibly nested) unsigned types

pub fn contains_optional_types(&self) -> bool

Check whether the interface contains any optional types

pub fn contains_sequence_types(&self) -> bool

Check whether the interface contains any sequence types

pub fn contains_map_types(&self) -> bool

Check whether the interface contains any map types

pub fn checksum(&self) -> u64

Calculate a numeric checksum for this ComponentInterface.

The checksum can be used to guard against accidentally using foreign-language bindings generated from one version of an interface with the compiled Rust code from a different version of that interface. It offers the following properties:

Two ComponentIntefaces generated from the same WebIDL file, using the same version of uniffi and the same version of Rust, will always have the same checksum value.

Two ComponentInterfaces will, with high probability, have different checksum values if: - They were generated from two different WebIDL files. - They were generated by two different versions of uniffi

The checksum may or may not change depending on the version of Rust used; since we expect consumers to be using the same executable to generate both the scaffolding and the bindings, assuming the same version of Rust seems acceptable.

Note that this is designed to prevent accidents, not attacks, so there is no need for the checksum to be cryptographically secure.

TODO: it’s not clear to me if the derivation of Hash is actually deterministic enough to ensure the guarantees above, or if it might be sensitive to e.g. compiler-driven re-ordering of struct field. Let’s see how it goes…

pub fn ffi_namespace(&self) -> String

The namespace to use in FFI-level function definitions.

The value returned by this method is used as a prefix to namespace all FFI-level functions used in this ComponentInterface.

Since these names are an internal implementation detail that is not typically visible to consumers, we take the opportunity to add an additional safety guard by including a 4-hex-char checksum in each name. If foreign-language bindings attempt to load and use a version of the Rust code compiled from a different UDL definition than the one used for the bindings themselves, then there is a high probability of checksum mismatch and they will fail to link against the compiled Rust code. The result will be an ugly inscrutable link-time error, but that is a lot better than triggering potentially arbitrary memory unsafety!

pub fn ffi_rustbuffer_alloc(&self) -> FFIFunction

Builtin FFI function for allocating a new RustBuffer. This is needed so that the foreign language bindings can create buffers in which to pass complex data types across the FFI.

pub fn ffi_rustbuffer_from_bytes(&self) -> FFIFunction

Builtin FFI function for copying foreign-owned bytes This is needed so that the foreign language bindings can create buffers in which to pass complex data types across the FFI.

pub fn ffi_rustbuffer_free(&self) -> FFIFunction

Builtin FFI function for freeing a RustBuffer. This is needed so that the foreign language bindings can free buffers in which they received complex data types returned across the FFI.

pub fn ffi_rustbuffer_reserve(&self) -> FFIFunction

Builtin FFI function for reserving extra space in a RustBuffer. This is needed so that the foreign language bindings can grow buffers used for passing complex data types across the FFI.

pub fn iter_ffi_function_definitions(&self) -> Vec<FFIFunction>

List the definitions of all FFI functions in the interface.

The set of FFI functions is derived automatically from the set of higher-level types along with the builtin FFI helper functions.

pub fn iter_rust_buffer_ffi_function_definitions(&self) -> Vec<FFIFunction>

fn resolve_type_expression<T: TypeResolver>(&mut self, expr: T) -> Result<Type>

Resolve a weedle type expression into a Type.

This method uses the current state of our TypeUniverse to turn a weedle type expression into a concrete Type (or error if the type expression is not well defined). It abstracts away the complexity of walking weedle’s type struct heirarchy by dispatching to the TypeResolver trait.

fn resolve_return_type_expression(
&mut self,
expr: &ReturnType<'_>
) -> Result<Option<Type>>

Resolve a weedle ReturnType expression into an optional Type.

This method is similar to resolve_type_expression, but tailored specifically for return types. It can return None to represent a non-existent return value.

fn add_namespace_definition(&mut self, defn: Namespace) -> Result<()>

Called by APIBuilder impls to add a newly-parsed namespace definition to the ComponentInterface.

fn add_enum_definition(&mut self, defn: Enum)

Called by APIBuilder impls to add a newly-parsed enum definition to the ComponentInterface.

fn add_record_definition(&mut self, defn: Record)

Called by APIBuilder impls to add a newly-parsed record definition to the ComponentInterface.

fn add_function_definition(&mut self, defn: Function) -> Result<()>

Called by APIBuilder impls to add a newly-parsed function definition to the ComponentInterface.

fn add_object_definition(&mut self, defn: Object)

Called by APIBuilder impls to add a newly-parsed object definition to the ComponentInterface.

fn add_callback_interface_definition(&mut self, defn: CallbackInterface)

Called by APIBuilder impls to add a newly-parsed callback interface definition to the ComponentInterface.

fn add_error_definition(&mut self, defn: Error)

Called by APIBuilder impls to add a newly-parsed error definition to the ComponentInterface.

fn check_consistency(&self) -> Result<()>

Perform global consistency checks on the declared interface.

This method checks for consistency problems in the declared interface as a whole, and which can only be detected after we’ve finished defining the entire interface.

fn derive_ffi_funcs(&mut self) -> Result<()>

Automatically derive the low-level FFI functions from the high-level types in the interface.

This should only be called after the high-level types have been completed defined, otherwise the resulting set will be missing some entries.