Trait uniffi_bindgen::interface::APIBuilder

trait APIBuilder {
    // Required method
    fn process(&self, ci: &mut ComponentInterface) -> Result<()>;
}

Trait to help build a ComponentInterface from WedIDL syntax nodes.

This trait does structural matching on the various weedle AST nodes and uses them to build up the records, enums, objects etc in the provided ComponentInterface.

Required Methods

fn process(&self, ci: &mut ComponentInterface) -> Result<()>