Trait uniffi_bindgen::interface::APIConverter

trait APIConverter<T> {
    fn convert(&self, ci: &mut ComponentInterface) -> Result<T>;
}

Trait to help convert WedIDL syntax nodes into ComponentInterface objects.

This trait does structural matching on the various weedle AST nodes and converts them into appropriate structs that we can use to build up the contents of a ComponentInterface. It is basically the TryFrom trait except that the conversion always happens in the context of a given ComponentInterface, which is used for resolving e.g. type definitions.

The difference between this trait and APIBuilder is that APIConverter treats the ComponentInterface as a read-only data source for resolving types, while APIBuilder actually mutates the ComponentInterface to add new definitions.

Required Methods

fn convert(&self, ci: &mut ComponentInterface) -> Result<T>