Module uniffi_bindgen::interface::types::resolver

Helpers for resolving UDL type expressions into concrete types.

This module provides the TypeResolver trait, an abstraction for walking the parse tree of a weedle type expression and using a TypeUniverse to convert it into a concrete type definition (so it assumes that you’re already used a TypeFinder to populate the universe).

Perhaps most importantly, it knows how to error out if the UDL tries to reference an undefined or invalid type.

Traits

TypeResolver

Trait to help resolving an UDL type node to a Type.

Functions

resolve_builtin_type

Resolve built-in API types by name.

validate_record_key_type

Validate that the type can be used for record keys.