Trait uniffi_bindgen::interface::types::resolver::TypeResolver

pub(crate) trait TypeResolver {
    // Required method
    fn resolve_type_expression(&self, types: &mut TypeUniverse) -> Result<Type>;
}

Trait to help resolving an UDL type node to a Type.

This trait does structural matching against type-related weedle AST nodes from a parsed UDL file, turning them into a corresponding Type struct. It uses the known type definitions in a TypeUniverse to resolve names to types.

As a side-effect, resolving a type expression will grow the type universe with references to the types seen during traversal. For example resolving the type expression sequence<<TestRecord>? will:

add Optional<Sequence<TestRecord> and Sequence<TestRecord> to the known types in the universe.

error out if the type name TestRecord is not already known.

Required Methods

fn resolve_type_expression(&self, types: &mut TypeUniverse) -> Result<Type>