Struct uniffi_bindgen::interface::types::TypeUniverse

pub(crate) struct TypeUniverse {
    type_definitions: HashMap<String, Type>,
    all_known_types: BTreeSet<Type>,
}

The set of all possible types used in a particular component interface.

Every component API uses a finite number of types, including primitive types, API-defined types like records and enums, and recursive types such as sequences of the above. Our component API doesn’t support fancy generics so this is a finitely-enumerable set, which is useful to be able to operate on explicitly.

You could imagine this struct doing some clever interning of names and so-on in future, to reduce the overhead of passing around Type instances. For now we just do a whole lot of cloning.

Fields
type_definitions: HashMap<String, Type>all_known_types: BTreeSet<Type>
Implementations

impl TypeUniverse

pub(super) fn add_type_definitions_from<T: TypeFinder>(
&mut self,
defn: T
) -> Result<()>

Add the definitions of all named Types from a given WebIDL definition.

This will fail if you try to add a name for which an existing type definition exists.

pub fn add_type_definition(&mut self, name: &str, type_: Type) -> Result<()>

Add the definition of a named Type.

This will fail if you try to add a name for which an existing type definition exists.

pub(super) fn get_type_definition(&self, name: &str) -> Option<Type>

Get the Type corresponding to a given name, if any.

pub(crate) fn resolve_type_expression<T: TypeResolver>(
&mut self,
expr: T
) -> Result<Type>

Get the Type corresponding to a given WebIDL type node.

If the node is a structural type (e.g. a sequence) then this will also add it to the set of all types seen in the component interface.

pub fn add_known_type(&mut self, type_: Type) -> Result<Type>

Add a Type to the set of all types seen in the component interface.

This helpfully returns a Result<Type> so it can be chained in with other methods during the type resolution process.

pub fn iter_known_types(&self) -> impl Iterator<Item = Type> + '_

Iterator over all the known types in this universe.