Trait uniffi_bindgen::interface::types::IterTypes

pub trait IterTypes {
    fn iter_types(&self) -> TypeIterator<'_>;
}

A trait for objects that may contain references to types.

Various objects in our interface will contain (possibly nested) references to types - for example a Record struct will contain one or more Field structs which will each have an associated type. This trait provides a uniform interface for inspecting the types references by an object.

Required methods

fn iter_types(&self) -> TypeIterator<'_>

Iterate over all types contained within on object.

This method iterates over the types contained with in object, making no particular guarantees about ordering or handling of duplicates.

The return type is a Box in order to hide the concrete implementation details of the iterator. Ideally we would return impl Iterator here but that’s not currently supported for trait methods.

Loading content...