Trait uniffi_bindgen::BindingGenerator

pub trait BindingGenerator: Sized {
    type Config: BindingsConfig;

    // Required method
    fn write_bindings(
        &self,
        ci: ComponentInterface,
        config: Self::Config,
        out_dir: &Utf8Path
    ) -> Result<()>;
}

A trait representing a UniFFI Binding Generator

External crates that implement binding generators, should implement this type and call the generate_external_bindings using a type that implements this trait.

Required Associated Types

type Config: BindingsConfig

Handles configuring the bindings

Required Methods

fn write_bindings( &self, ci: ComponentInterface, config: Self::Config, out_dir: &Utf8Path ) -> Result<()>

Writes the bindings to the output directory

Arguments

ci: A ComponentInterface representing the interface

config: An instance of the BindingsConfig associated with this type

out_dir: The path to where the binding generator should write the output bindings